# backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "FORJA API",
    "version": "0.1.0",
    "description": "Generado desde los schemas zod del backend (backend/src/openapi). No editar a mano — correr `npm run openapi:generate` después de cambiar una ruta."
  },
  "servers": [
    {
      "url": "http://localhost:4000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Token de sesión emitido por /auth/login o /auth/register."
      },
      "clientIdLegacy": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Client-Id",
        "description": "Modo legacy sin cuenta: identifica la instalación. Se ignora si viene Authorization Bearer válido."
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "nullable": true
          }
        }
      },
      "OkResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ]
      },
      "WorkoutType": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "muscle_group": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "label",
          "muscle_group"
        ]
      },
      "ExerciseInfo": {
        "type": "object",
        "properties": {
          "exercise_name": {
            "type": "string"
          },
          "default_rest_seconds": {
            "type": "number"
          },
          "target_sets": {
            "type": "number",
            "nullable": true
          },
          "target_reps": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "exercise_name",
          "default_rest_seconds",
          "target_sets",
          "target_reps"
        ]
      },
      "ExerciseUpsertBody": {
        "type": "object",
        "properties": {
          "exercise_name": {
            "type": "string",
            "minLength": 1
          },
          "target_sets": {
            "type": "number",
            "nullable": true
          },
          "target_reps": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "exercise_name"
        ]
      },
      "ExercisePatchBody": {
        "type": "object",
        "properties": {
          "target_sets": {
            "type": "number",
            "nullable": true
          },
          "target_reps": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "target_sets",
          "target_reps"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/api/workout-types": {
      "get": {
        "tags": [
          "workout-types"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "clientIdLegacy": []
          }
        ],
        "responses": {
          "200": {
            "description": "Catálogo de tipos de entreno",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkoutType"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error de validación",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Falta identidad (token inválido o X-Client-Id ausente)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/workout-types/{id}/exercises": {
      "get": {
        "tags": [
          "workout-types"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "clientIdLegacy": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Ejercicios predeterminados de ese tipo",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExerciseInfo"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error de validación",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Falta identidad (token inválido o X-Client-Id ausente)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Tipo no encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "workout-types"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "clientIdLegacy": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExerciseUpsertBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Agregado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error de validación",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Falta identidad (token inválido o X-Client-Id ausente)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Tipo no encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/workout-types/{id}/exercises/{exerciseName}": {
      "patch": {
        "tags": [
          "workout-types"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "clientIdLegacy": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "exerciseName",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExercisePatchBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Actualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error de validación",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Falta identidad (token inválido o X-Client-Id ausente)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "workout-types"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "clientIdLegacy": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "exerciseName",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Eliminado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error de validación",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Falta identidad (token inválido o X-Client-Id ausente)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}